+++ src/main/java/delta/games/lotro/tools/dat/emotes/MainDatEmotesLoader.java
@@ -73,7 +73,7 @@ Emote_SourceText:
   private EmoteDescription load(int indexDataId)
   {
     EmoteDescription emote=null;
-    int dbPropertiesId=indexDataId+DATConstants.DBPROPERTIES_OFFSET;
+    long dbPropertiesId=indexDataId+DATConstants.DBPROPERTIES_OFFSET;
     PropertiesSet properties=_facade.loadProperties(dbPropertiesId);
     if (properties!=null)
     {
