+++ src/main/java/delta/games/lotro/tools/dat/items/FieldIconsLoader.java
@@ -37,7 +37,7 @@ public class FieldIconsLoader
 
   private void handleRecipe(int indexDataId)
   {
-    int dbPropertiesId=indexDataId+DATConstants.DBPROPERTIES_OFFSET;
+    long dbPropertiesId=indexDataId+DATConstants.DBPROPERTIES_OFFSET;
     PropertiesSet properties=_facade.loadProperties(dbPropertiesId);
     if (properties!=null)
     {
